Use basename(3p) to fix copying to directory
Previously the full (or relative) path of the source was used when
copying into a directory. This resulted in the path of the source being
appended to the path of the destination which could place the file
somewhere unexpected.

E.g. `./tcp ../file .` resulted in `file` being copied to `./../file`
which puts it back where it started, not at all expected behavior.