Add recursive mode
Added ftscmp.c and recursive_traverse.c in order to use the fts(3)
family of functions for the recursive traverse mode. Need to change some
error messages to be non-fatal (permission errors for example).

Removed swap.c, it is unneeded.

Added common includes to the header file (also needed for definition of
ftscmp.c).