Fix realloc bug in tokenizePipeline
The function was not reallocating space for consecutive words in the
same pipeline segment.

In order to fix this the word realloc call was moved to outside of the
check for whether or not the token is a pipe ('|') or a word.

Also minor update to the trace output to get rid of trailing space.