Implement basic I/O redirection
This saves a backup of the default stdin and stdout file descriptors at
the start of execution, then will change the file descriptors when a
control sequence is found and restore them after the command has
finished executing.

This works for builtins (although that doesn't mean much except for
echo). The I/O redirection 'words' need to be removed from the pipeline
structure in order to work with commands.