Fix I/O redirection for commands
The control 'words' had been left in the command string and passed to
exec causing commands to not be executed properly.

This adds 'removeIndex' and removes the I/O redirection control 'words'
from the command string before passing the command to exec.

Still need to check for illegal I/O redirection combinations.