Fix I/O redirection append errors
First append didn't work. To fix this the check for append was moved to
before the check for redirecting stdout (because their control
characters start the same).

Then append opened an invalid file descriptor. To fix this the
'O_WRONLY' flag was added, because 'O_APPEND' doesn't actually specify
how to open the file.

The it was found that redirecting stdout also appended. To fix this the
'O_TRUNC' flag was added.

Also 'restoreIO' now closes file descriptors before restoring them from
the saved backups.